backend/requirements.txt
fastapi==0.136.1
uvicorn==0.27.0
pytest==9.0.3
httpx==0.27.0
pydantic-settings==2.1.0
aiosmtplib==3.0.1
aioimaplib==1.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
pytest-asyncio
openai==1.12.0
langchain-text-splitters==1.1.2
tiktoken==0.6.0
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
email-validator==2.1.0
cryptography==47.0.0
PyJWT==2.10.1
python-multipart==0.0.27
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
setuptools==78.1.1
